requirements.txt
streamlit==1.38.0 
langchain-community==0.2.16 
langgraph==0.2.14 
fpdf==1.7.2 
python-dotenv==1.0.1
requests==2.31.0
fastapi==0.115.0
uvicorn==0.30.6